editoast: move train schedule path item to schemas
