+++ editoast/src/views/v2/path/pathfinding.rs
@@ -11,6 +11,7 @@ use diesel_async::AsyncPgConnection as PgConnection;
 use editoast_common::Identifier;
 use editoast_schemas::rolling_stock::LoadingGaugeType;
 use editoast_schemas::track_offset::TrackOffset;
+use editoast_schemas::train_schedule::PathItemLocation;
 use serde::Deserialize;
 use serde::Serialize;
 use utoipa::ToSchema;
@@ -28,7 +29,6 @@ use crate::modelsv2::TrackSectionModel;
 use crate::redis_utils::RedisClient;
 use crate::redis_utils::RedisConnection;
 use crate::schema::operational_point::OperationalPoint;
-use crate::schema::v2::trainschedule::PathItemLocation;
 use crate::views::get_app_version;
 use crate::views::v2::path::PathfindingError;
 use crate::views::v2::path::TrackRange;
